+++ tree/math_tree/math_tree_node_data.h
@@ -16,7 +16,7 @@ enum MathNodeOperator {
 
 union MathNodeData {
 
-    int mathNodeValue;
+    double mathNodeValue;
     MathNodeOperator mathOperator;
 };
